requirements-app.txt
discord.py==2.3.2
Django==4.2.5
openai==0.27.1
langchain==0.0.301
langchainplus-sdk==0.0.20
tiktoken==0.2.0
python-dotenv==1.0.0
weaviate-client==3.24.1
gunicorn==21.2.0